requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0
duckduckgo-search>=3.0.0
mediapipe>=0.10.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
google-cloud-storage>=2.0.0 